src/requirements.txt
appdirs
CacheControl
certifi
cffi
chardet
colorama
contextlib2
Cython
decorator
distlib
distro
filelock
html5lib
idna
JACK-Client
alsa-midi
mididings
msgpack
ordered-set
packaging
pep517
Pillow
ply
progress
psutil
pycparser
pyliblo
pyparsing
pyserial
python-rtmidi
pytoml
requests
retrying
six
toml
tornado
urllib3
virtualenv
webencodings
